matching service update

+++ user-service/src/main/java/com/mood/userservice/vo/MatchingData.java
@@ -3,6 +3,6 @@ package com.mood.userservice.vo;
 import java.time.LocalDateTime;
 
 public class MatchingData {
-    private int moodDistance;
+    private double moodDistance;
     private LocalDateTime matchingTime;
 }
